tests/wpt/web-platform-tests/html/rendering/non-replaced-elements/tables/table-vspace-hspace.html

<!-- quirks -->
<meta charset=utf-8>
<title>table vspace hspace (quirks mode)</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div>x</div> <!-- prevent margin collapsing quirks -->
<table vspace=25 hspace=25><tr><td>x</table>
<div>x</div> <!-- prevent margin collapsing quirks -->
<script>
var style = getComputedStyle(document.querySelector('table'));
['marginTop', 'marginRight', 'marginBottom', 'marginLeft'].forEach(function(m) {
  assert_equals(style[m], '0px', m);
});
done();
</script>
