LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/tables/table-vspace-hspace-expected.txt

CONSOLE MESSAGE: line 3463: Error: assert_equals: marginTop expected "0px" but got "25px"
x
x
x

Harness Error (FAIL), message = Error: assert_equals: marginTop expected "0px" but got "25px"
